requirements.txt
PySide2
iconify
mysql-connector
